Fix/hf (#392)
* Use _id instead, as `id` is subject to change

The `id` of a dataset can change when a user changes its name or
the datasets' name. The `_id` is persistent, which means that we
can keep track of whether or not the dataset has already been
indexed. If any part of the name changes, the old `name` simply
redirects. In the worst case (bare _id becoming invalid), we could
always reindex and match based on `_id`.

* Migration script for using stable _id instead of id for HF

* Skip datassets that have already been updated

* Fix reference to wrong field for platform_resource_identifier

* Make sure triggers get added on table creation, avoid hack

* Make sure deletion triggers also added to existing database

* Black formatting

* Add comment referencing the PR/issue for documentation